all current issues solved
