+++ com_migratetojoomla/admin/forms/parameter.xml
@@ -7,8 +7,8 @@
 		   default="1"
 		   layout="joomla.form.field.radio.switcher"
 		   validate="options">
-		<option value="1">JYES</option>
 		<option value="0">JNO</option>
+		<option value="1">JYES</option>
 	</field>
 	<field name="databasemigratestatus"
 		   type="radio"
@@ -17,7 +17,7 @@
 		   default="1"
 		   layout="joomla.form.field.radio.switcher"
 		   validate="options">
-		<option value="1">JYES</option>
 		<option value="0">JNO</option>
+		<option value="1">JYES</option>
 	</field>
 </form>
