+++ com_migratetojoomla/admin/language/en-GB/com_migratetojoomla.sys.ini
@@ -1,5 +1,5 @@
 ; Joomla! Project
-; (C) 2010 Open Source Matters, Inc. <https://www.joomla.org>
+; (C) 2024 Open Source Matters, Inc. <https://www.joomla.org>
 ; License GNU General Public License version 2 or later; see LICENSE.txt
 ; Note : All ini files need to be saved as UTF-8
 
@@ -38,7 +38,7 @@ COM_MIGRATETOJOOMLA_FS="File System"
 COM_MIGRATETOJOOMLA_FS_CONNECTION_SUCCESSFULLY="File system Connection Successfully"
 COM_MIGRATETOJOOMLA_FS_CONNECTION_UNSUCCESSFULLY="File system Connection Unuccessfully"
 COM_MIGRATETOJOOMLA_FTP="FTP"
-COM_MIGRATETOJOOMLA_FTP_BASE_DIR_DESC="Path of FTP base directory where wordpress install"
+COM_MIGRATETOJOOMLA_FTP_BASE_DIR_DESC="Path of FTP base directory where WordPress install"
 COM_MIGRATETOJOOMLA_FTP_BASE_DIR_LABEL="FTP base directory"
 COM_MIGRATETOJOOMLA_FTP_CONNECTION_SUCCESFULLY="Ftp Connection Successfully"
 COM_MIGRATETOJOOMLA_FTP_CONNECTION_UNSUCCESFULLY="Ftp Connection Unsuccessfully"
@@ -59,7 +59,7 @@ COM_MIGRATETOJOOMLA_HTTP="Http"
 COM_MIGRATETOJOOMLA_HTTP_CONNECTION_SUCCESSFULLY="Http Connection Successfully"
 COM_MIGRATETOJOOMLA_HTTP_CONNECTION_UNSUCCESSFULLY="Http Connection Unsuccessfuly"
 COM_MIGRATETOJOOMLA_HTTP_LINK_DESC="Link of Host website (Make sure Directory listing Allow on server)"
-COM_MIGRATETOJOOMLA_HTTP_LINK_LABEL="Wordpress live website URL"
+COM_MIGRATETOJOOMLA_HTTP_LINK_LABEL="WordPress live website URL"
 COM_MIGRATETOJOOMLA_IMPORT_SUCCESSFULLY = "Import SuccessFully"
 COM_MIGRATETOJOOMLA_IMPORT_UNSUCCESSFULLY = "Import UnsuccessFully"
 COM_MIGRATETOJOOMLA_MEDIA_CONNECTION_SUCCESSFULLY="Media Connection Successfully"
@@ -89,7 +89,7 @@ COM_MIGRATETOJOOMLA_WEBSITE_BASE_DIRECTORY_DESC="Base Directory Path"
 COM_MIGRATETOJOOMLA_WEBSITE_BASE_DIRECTORY_LABEL="Website base directory"
 COM_MIGRATETOJOOMLA_WILL_MIGRATE="Will be migrated"
 COM_MIGRATETOJOOMLA_WILL_NOT_MIGRATE="Will not be migrated"
-COM_MIGRATETOJOOMLA_WORDPRESS="Wordpress"
-COM_MIGRATETOJOOMLA_WORDPRESS_FTP_PARAMETERS="Wordpress ftp parameters"
-COM_MIGRATETOJOOMLA_WORDPRESS_WEBSITE_PARAMETERS="Wordpress Website Parameters"
+COM_MIGRATETOJOOMLA_WORDPRESS="WordPress"
+COM_MIGRATETOJOOMLA_WORDPRESS_FTP_PARAMETERS="WordPress ftp parameters"
+COM_MIGRATETOJOOMLA_WORDPRESS_WEBSITE_PARAMETERS="WordPress Website Parameters"
 COM_MIGRATETOJOOMLA_XML_DESCRIPTION="This component migrates websites from other platforms to Joomla."
