+++ com_migratetojoomla/admin/services/provider.php
@@ -4,7 +4,7 @@
  * @package     Joomla.Administrator
  * @subpackage  com_migratetojoomla
  *
- * @copyright   (C) 2018 Open Source Matters, Inc. <https://www.joomla.org>
+ * @copyright   (C) 2024 Open Source Matters, Inc. <https://www.joomla.org>
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
 
@@ -28,7 +28,7 @@ use Joomla\DI\ServiceProviderInterface;
  */
 
 return new class implements ServiceProviderInterface {
-    
+
     /**
      * Registers the service provider with a DI container.
      *
