+++ com_migratetojoomla/admin/src/controller/CheckController.php
@@ -4,7 +4,7 @@
  * @package     Joomla.Administrator
  * @subpackage  com_migratetojoomla
  *
- * @copyright   (C) 2006 Open Source Matters, Inc. <https://www.joomla.org>
+ * @copyright   (C) 2024 Open Source Matters, Inc. <https://www.joomla.org>
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
 
@@ -15,10 +15,8 @@ namespace Joomla\Component\MigrateToJoomla\Administrator\Controller;
 // phpcs:enable PSR1.Files.SideEffects
 
 /**
- * Migrate controller class.
+ * Check controller class.
  *
  * @since  1.0
  */
-class CheckController 
-{
-}
+class CheckController {}
