+++ com_migratetojoomla/admin/src/controller/DisplayController.php
@@ -4,8 +4,8 @@
  * @package     Joomla.Administrator
  * @subpackage  com_migratetojoomla
  *
- * @copyright     (C) 2023 Open Source Matters, Inc. <https://www.joomla.org>
- * @license       GNU General Public License version 2 or later; see LICENSE.txt
+ * @copyright   (C) 2024 Open Source Matters, Inc. <https://www.joomla.org>
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
 
 namespace Joomla\Component\MigrateToJoomla\Administrator\Controller;
@@ -27,7 +27,7 @@ use Joomla\CMS\MVC\Controller\BaseController;
 class DisplayController extends BaseController
 {
     /**
-     * The default view for the display method.
+     * Default view
      *
      * @var string
      */
