+++ com_migratetojoomla/admin/src/controller/ProgressController.php
@@ -4,7 +4,7 @@
  * @package     Joomla.Administrator
  * @subpackage  com_migratetojoomla
  *
- * @copyright   (C) 2006 Open Source Matters, Inc. <https://www.joomla.org>
+ * @copyright   (C) 2024 Open Source Matters, Inc. <https://www.joomla.org>
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
 
@@ -23,7 +23,7 @@ use Joomla\Component\MigrateToJoomla\Administrator\Helper\LogHelper;
 // phpcs:enable PSR1.Files.SideEffects
 
 /**
- * Migrate controller class.
+ * Progress controller class.
  *
  * @since  1.0
  */
