+++ com_migratetojoomla/admin/tmpl/information/default.php
@@ -4,7 +4,7 @@
  * @package     Joomla.Administrator
  * @subpackage  com_migratetojoomla
  *
- * @copyright     (C) 2023 Open Source Matters, Inc. <https://www.joomla.org>
+ * @copyright     (C) 2024 Open Source Matters, Inc. <https://www.joomla.org>
  * @license       GNU General Public License version 2 or later; see LICENSE.txt
  */
