+++ com_migratetojoomla/media/js/admin-migratetojoomla.js
@@ -1,5 +1,5 @@
 /**
- * @copyright  (C) 2018 Open Source Matters, Inc. <https://www.joomla.org>
+ * @copyright  (C) 2024 Open Source Matters, Inc. <https://www.joomla.org>
  * @license    GNU General Public License version 2 or later; see LICENSE.txt
  */
