+++ com_migratetojoomla/migratetojoomla.xml
@@ -2,12 +2,12 @@
 <extension type="component" method="upgrade">
     <name>MigrateToJoomla</name>
     <author>Joomla! Project</author>
-	<creationDate>2023</creationDate>
-	<copyright>(C) 2013 Open Source Matters, Inc.</copyright>
-	<license>GNU General Public License version 2 or later; see LICENSE.txt</license>
-	<authorEmail>admin@joomla.org</authorEmail>
-	<authorUrl>www.joomla.org</authorUrl>
-	<version>1.0</version>
+    <creationDate>2024</creationDate>
+    <copyright>(C) 2024 Open Source Matters, Inc.</copyright>
+    <license>GNU General Public License version 2 or later; see LICENSE.txt</license>
+    <authorEmail>admin@joomla.org</authorEmail>
+    <authorUrl>www.joomla.org</authorUrl>
+    <version>1.0</version>
     <description>COM_MIGRATETOJOOMLA_XML_DESCRIPTION</description>
     <namespace path="src">Joomla\Component\MigrateToJoomla</namespace>
     <administration>
@@ -23,11 +23,11 @@
         <languages folder="admin">
             <language tag="en-GB">language/en-GB/com_migratetojoomla.ini</language>
             <language tag="en-GB">language/en-GB/com_migratetojoomla.sys.ini</language>
-		</languages>
+        </languages>
     </administration>
     <media destination="com_migratetojoomla" folder="media">
-		<folder>js</folder>
-		<folder>css</folder>
+        <folder>js</folder>
+        <folder>css</folder>
         <file>joomla.asset.json</file>
-	</media>
+    </media>
 </extension>
