+++ pkg_migratetojoomla/languages/en-GB/pkg_migratetojoomla.ini
@@ -1,0 +1,2 @@
+PKG_MIGRATETOJOOMLA="Migrate To Joomla Package"
+PKG_MIGRATETOJOOMLA_XML_DESCRIPTION="Migrate To Joomla ! Migrate Website From other Platforms to Joomla."
