+++ plugins/plg_migratetojoomla_mediadownload/language/en-GB/plg_migratetojoomla_mediadownload.ini
@@ -1,5 +1,5 @@
 ; Joomla! Project
-; (C) 2017 Open Source Matters, Inc. <https://www.joomla.org>
+; (C) 2024 Open Source Matters, Inc. <https://www.joomla.org>
 ; License GNU General Public License version 2 or later; see LICENSE.txt
 ; Note : All ini files need to be saved as UTF-8
