+++ plugins/plg_migratetojoomla_mediadownload/mediadownload.xml
@@ -2,8 +2,8 @@
 <extension type="plugin" group="migratetojoomla" method="upgrade">
 	<name>PLG_MIGRATETOJOOMLA_MEDIADOWNLOAD</name>
     <author>Joomla! Project</author>
-	<creationDate>2023</creationDate>
-	<copyright>(C) 2013 Open Source Matters, Inc.</copyright>
+	<creationDate>2024</creationDate>
+	<copyright>(C) 2024 Open Source Matters, Inc.</copyright>
 	<license>GNU General Public License version 2 or later; see LICENSE.txt</license>
 	<authorEmail>admin@joomla.org</authorEmail>
 	<authorUrl>www.joomla.org</authorUrl>
