+++ plugins/plg_migratetojoomla_mediadownload/src/Extension/filesystem.php
@@ -2,9 +2,9 @@
 
 /**
  * @package     Joomla.Administrator
- * @subpackage  com_migratetojoomla
+ * @subpackage  com_migratetojoomla.mediadownload
  *
- * @copyright   (C) 2006 Open Source Matters, Inc. <https://www.joomla.org>
+ * @copyright   (C) 2024 Open Source Matters, Inc. <https://www.joomla.org>
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
