+++ plugins/plg_migratetojoomla_wordpress/language/en-GB/plg_migratetojoomla_wordpress.sys.ini
@@ -1,11 +1,11 @@
 ; Joomla! Project
-; (C) 2017 Open Source Matters, Inc. <https://www.joomla.org>
+; (C) 2024 Open Source Matters, Inc. <https://www.joomla.org>
 ; License GNU General Public License version 2 or later; see LICENSE.txt
 ; Note : All ini files need to be saved as UTF-8
 
 PLG_MIGRATETOJOOMLA_WORDPRESS_ADMINISTRATOR="Administrator"
 PLG_MIGRATETOJOOMLA_WORDPRESS_AUTHOR="Author"
-PLG_MIGRATETOJOOMLA_WORDPRESS="Wordpress"
+PLG_MIGRATETOJOOMLA_WORDPRESS="WordPress"
 PLG_MIGRATETOJOOMLA_WORDPRESS_DESC="Plugin that extend migration of wordpress with com_migratetojoomla"
 PLG_MIGRATETOJOOMLA_WORDPRESS_EDITOR="Editor"
 PLG_MIGRATETOJOOMLA_WORDPRESS_FEATURE_IMAGE_AS_BOTH_FULL_INTRO="As Artcile Full and Intro image"
