+++ plugins/plg_migratetojoomla_wordpress/src/Extension/Wordpress.php
@@ -1,10 +1,10 @@
 <?php
 
 /**
- * @package     Joomla.Plugin
- * @subpackage  Content.contact
+ * @package     Joomla.Administrator
+ * @subpackage  com_migratetojoomla.wordpress
  *
- * @copyright   (C) 2014 Open Source Matters, Inc. <https://www.joomla.org>
+ * @copyright   (C) 2024 Open Source Matters, Inc. <https://www.joomla.org>
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
